backend/pyproject.toml
[project]
name = "documentation-mcp-backend"
version = "0.1.0"
description = "Backend infrastructure for Documentation MCP Gateway"
requires-python = ">=3.13"
dependencies = [
  "alembic>=1.16.5",
  "celery[redis]>=5.5.3",
  "crawl4ai>=0.8.0,<0.9.0",
  "fastapi>=0.121.0",
  "pgvector>=0.4.1",
  "psycopg[binary]>=3.2.13",
  "pydantic-settings>=2.11.0",
  "redis>=5.2.1,<6.5",
  "sqlmodel>=0.0.27",
  "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
  "httpx>=0.28.1",
  "pytest>=8.4.2",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["."]

[tool.uv]
package = false

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"
